Module mortal::terminal

Provides an interface to terminal devices

Structs

Cursor

Represents the cursor position in a terminal device

ModifierState

Represents a set of modifier keys

MouseEvent

Represents a mouse event

PrepareConfig

Configures a Terminal or Screen instance to read special input.

PrepareState

Represents a previous device state of a Terminal.

Size

Represents the size of a terminal window

Style

Represents a set of style attributes applied to text.

Terminal

Provides concurrent read and write access to a terminal device

TerminalReadGuard

Holds an exclusive lock for read operations on a Terminal

TerminalWriteGuard

Holds an exclusive lock for write operations on a Terminal

Enums

Color

Represents a color attribute applied to text foreground or background.

CursorMode

Represents the visual appearance of the cursor in the terminal

Event

Represents an event generated from a terminal interface

Key

Represents a keyboard key press event

MouseButton

Represents a button on a mouse device

MouseInput

Represents the type of mouse input event